+++ include/rain/algorithm.hpp
@@ -2,6 +2,7 @@
 #pragma once
 
 #include "algorithm/algorithm.hpp"
+#include "algorithm/bigint.hpp"
 #include "algorithm/bit-manipulators.hpp"
 #include "algorithm/dsu.hpp"
 #include "algorithm/edmonds-karp.hpp"
@@ -10,5 +11,7 @@
 #include "algorithm/huffman.hpp"
 #include "algorithm/kmp.hpp"
 #include "algorithm/lru.hpp"
-#include "algorithm/prime-modulus-field.hpp"
+#include "algorithm/miller-rabin.hpp"
+#include "algorithm/modulus-field.hpp"
 #include "algorithm/segment-tree.hpp"
+#include "algorithm/sieve.hpp"